// API.ts
export type Todo = {
  __typename: "Todo",
  content: string,
  createdAt: string,
  done: boolean,
  id: string,
  owner?: string | null,
  priority?: TodoPriority | null,
  updatedAt: string,
};

export enum TodoPriority {
  high = "high",
  low = "low",
  medium = "medium",
}
// ...
export type ModelTodoConnection = {
  __typename: "ModelTodoConnection",
  items:  Array<Todo | null >,
  nextToken?: string | null,
};
// ...
export type ListTodosQuery = {
  listTodos?:  {
    __typename: "ModelTodoConnection",
    items:  Array< {
      __typename: "Todo",
      content: string,
      createdAt: string,
      done: boolean,
      id: string,
      owner?: string | null,
      priority?: TodoPriority | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};